media: i2c: Add activity indicator for OmniVision OV02C10
CSI MIPI cameras do not natively support activity indicator, as it is
typically an add-on GPIO-controlled feature. If such indicator is
present, for privacy reasons, it is desired to control it directly by
the driver, such that indicator cannot be easily disabled.

Add optional indicator pin, which can be enabled in device tree of
supported platforms. Just like UVC based webcam, turn on indicator on
stream-start, turn off on stream-stop.

Signed-off-by: Aleksandrs Vinarskis <alex.vinarskis@gmail.com>